+++ component/xload/block.go
@@ -31,7 +31,7 @@
    SOFTWARE
 */
 
-package common
+package xload
 
 import (
 	"fmt"
